Use bitmasking for tracking events.
Practically ignore one event from repeatedly tracked.
Though this will also ignore events order.

+++ event/notification.go
@@ -1,11 +1,11 @@
 package event
 
 type Intent struct {
-	EventId  byte
+	EventId  uint32
 	FileName string
 }
 
-func Notify(id byte, filename string) {
+func Notify(id uint32, filename string) {
 	handler.Chan <- &Intent{
 		EventId:  id,
 		FileName: filename,
